feat: controller 메소드 추가에 의한 수정

+++ src/main/java/com/devco/jpaproject/service/PostService.java
@@ -1,13 +1,16 @@
 package com.devco.jpaproject.service;
 
+import com.devco.jpaproject.controller.dto.PostDeleteRequestDto;
 import com.devco.jpaproject.controller.dto.PostRequestDto;
 import com.devco.jpaproject.controller.dto.PostResponseDto;
 import com.devco.jpaproject.domain.Post;
 import com.devco.jpaproject.domain.User;
+import com.devco.jpaproject.exception.PostNotFoundException;
+import com.devco.jpaproject.exception.UserAndPostNotMatchException;
+import com.devco.jpaproject.exception.UserNotFoundException;
 import com.devco.jpaproject.repository.PostRepository;
 import com.devco.jpaproject.repository.UserRepository;
 import com.devco.jpaproject.service.converter.Converter;
-import javassist.NotFoundException;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
@@ -27,29 +30,29 @@ public class PostService {
     private final Converter converter;
 
     @Transactional(readOnly = true)
-    public List<PostResponseDto> findAll(){
+    public List<PostResponseDto> findAll() {
         return postRepository.findAll().stream()
                 .map(converter::toPostResponseDto)
                 .collect(Collectors.toList());
     }
 
     @Transactional(readOnly = true)
-    public Page<PostResponseDto> findAllByPages(Pageable pageable){
+    public Page<PostResponseDto> findAllByPages(Pageable pageable) {
         return postRepository.findAll(pageable)
                 .map(converter::toPostResponseDto);
     }
 
     @Transactional(readOnly = true)
-    public PostResponseDto findById(Long id) throws NotFoundException {
+    public PostResponseDto findById(Long id) throws PostNotFoundException {
         return postRepository.findById(id)
                 .map(converter::toPostResponseDto)
-                .orElseThrow(() -> new NotFoundException("post not found"));
+                .orElseThrow(() -> new PostNotFoundException("postId: " + id + "not found"));
     }
 
     @Transactional
-    public Long insert(PostRequestDto requestDto) throws NotFoundException {
+    public Long insert(PostRequestDto requestDto) throws UserNotFoundException {
         User writer = userRepository.findById(requestDto.getWriterId())
-                .orElseThrow(() -> new NotFoundException("writer not found"));
+                .orElseThrow(() -> new UserNotFoundException("writerId: " + requestDto.getWriterId() + " not found"));
 
         Post post = converter.toPostEntity(requestDto, writer);
         postRepository.save(post);
@@ -58,9 +61,9 @@ public class PostService {
     }
 
     @Transactional
-    public void update(Long id, PostRequestDto requestDto) throws Exception {
+    public void update(Long id, PostRequestDto requestDto) throws PostNotFoundException {
         Post post = postRepository.findById(id)
-                .orElseThrow(() -> new Exception("post not found"));
+                .orElseThrow(() -> new PostNotFoundException("postId: " + id + "not found"));
 
         post.updateContent(requestDto.getContent());
         post.updateTitle(requestDto.getTitle());
@@ -68,16 +71,20 @@ public class PostService {
     }
 
     @Transactional
-    public void delete(Long id) throws Exception {
-        Post post = postRepository.findById(id)
-                .orElseThrow(() -> new Exception("post not found"));
+    public void deleteOne(PostDeleteRequestDto dto) throws PostNotFoundException, UserNotFoundException, UserAndPostNotMatchException {
+        Post post = postRepository.findById(dto.getPostId())
+                .orElseThrow(() -> new PostNotFoundException("postId: " + dto.getPostId() + "not found"));
+
+        User writer = userRepository.findById(dto.getWriterId())
+                .orElseThrow(() -> new UserNotFoundException("writerId: " + dto.getWriterId() + " not found"));
+
+        boolean result = writer.getPosts().stream()
+                .anyMatch(p -> p.getId() == post.getId());
+
+        if (!result)
+            throw new UserAndPostNotMatchException(String.format("postId: %d, writerId: %d", dto.getPostId(), dto.getWriterId()));
 
         post.delete();
         postRepository.delete(post);
-    }
-
-    @Transactional
-    public void deleteAll() {
-        postRepository.deleteAll();
     }
 }
